+++ test/unit/lib/device.js
@@ -88,21 +88,21 @@ describe('device', function () {
     });
   });
 
-  describe('isOldAndroidChromedBasedWebview()', function () {
+  describe('isKitKatWebview()', function () {
     it('returns true if user agent is Android 4 Webview with Chrome in useragent', function () {
-      expect(device.isOldAndroidChromeBasedWebview(AGENTS.androidWebviewKitKatLollipop)).to.equal(true);
+      expect(device.isKitKatWebview(AGENTS.androidWebviewKitKatLollipop)).to.equal(true);
     });
 
     it('returns false for Android webviews without chrome in user agent', function () {
-      expect(device.isOldAndroidChromeBasedWebview(AGENTS.androidWebviewOld)).to.equal(false);
+      expect(device.isKitKatWebview(AGENTS.androidWebviewOld)).to.equal(false);
     });
 
     it('returns false for Android webviews with newer builds of Chrome', function () {
-      expect(device.isOldAndroidChromeBasedWebview(AGENTS.androidWebviewLollipopAndAbove)).to.equal(false);
+      expect(device.isKitKatWebview(AGENTS.androidWebviewLollipopAndAbove)).to.equal(false);
     });
 
     it('returns false for Android Chrome', function () {
-      expect(device.isOldAndroidChromeBasedWebview(AGENTS.androidPhoneChrome)).to.equal(false);
+      expect(device.isKitKatWebview(AGENTS.androidPhoneChrome)).to.equal(false);
     });
   });