+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## UNRELEASED
 
-* Provide specific strategy for old Android webviews. Fixes issue where only 4 characters could be input
+* Provide specific strategy for Android KitKat Chromium based webviews. Fixes issue where only 4 characters could be input
 
 ## 1.0.11 (2017-02-22)